Add checkpoint for user-define storage; Add a mini-batch resource request method to avoid many PSs are allocated on a single node

+++ angel-ps/core/src/main/java/com/tencent/angel/ps/storage/vector/ServerRowFactory.java
@@ -134,7 +134,7 @@ public class ServerRowFactory {
             estEleNum);
 
       case T_ANY_LONGKEY_SPARSE:
-        return new ServerLongAnyRow(valueClass, rowIndex, rowType, (int) startCol, (int) endCol,
+        return new ServerLongAnyRow(valueClass, rowIndex, rowType, startCol, endCol,
             estEleNum);
 
       default:
